[Feat] #S10P31S105-284: api서버 지그 그래프 정보 반환 자료형 변경

+++ be/api-server/src/main/java/com/sdi/apiserver/api/jig/dto/response/JigGraphResponseDto.java
@@ -2,9 +2,9 @@ package com.sdi.apiserver.api.jig.dto.response;
 
 public record JigGraphResponseDto(
         int day,
-        float output,
-        float cost,
-        float yield,
+        double output,
+        double cost,
+        double yield,
         int countMissingJig
 ) {
 }
